+++ test/integration/dynamic_settings.sh
@@ -194,3 +194,38 @@ cat <<EOT > /etc/clickhouse-server/users.d/allow_deprecated_database_ordinary.xm
 EOT
 
 fi
+
+if [[ "${CLICKHOUSE_VERSION}" =~ ^20\.1[3-9] ]]; then
+
+cat <<EOT > /etc/clickhouse-server/users.d/allow_experimental_database_materialize_mysql.xml
+<yandex>
+<profiles><default>
+ <allow_experimental_database_materialize_mysql>1</allow_experimental_database_materialize_mysql>
+</default></profiles>
+</yandex>
+EOT
+
+fi
+
+if [[ "${CLICKHOUSE_VERSION}" == "head" || "${CLICKHOUSE_VERSION}" =~ ^2[1-9]\.[1-9] ]]; then
+cat <<EOT > /etc/clickhouse-server/users.d/allow_experimental_database_materialized_mysql.xml
+<yandex>
+<profiles><default>
+ <allow_experimental_database_materialized_mysql>1</allow_experimental_database_materialized_mysql>
+</default></profiles>
+</yandex>
+EOT
+
+fi
+
+if [[ "${CLICKHOUSE_VERSION}" == "head" || "${CLICKHOUSE_VERSION}" =~ ^2[2-9]\.[1-9] || "${CLICKHOUSE_VERSION}" =~ ^21\.[8-9] || "${CLICKHOUSE_VERSION}" =~ ^21\.1[0-9] ]]; then
+
+cat <<EOT > /etc/clickhouse-server/users.d/allow_experimental_database_materialized_postgresql.xml
+<yandex>
+<profiles><default>
+ <allow_experimental_database_materialized_postgresql>1</allow_experimental_database_materialized_postgresql>
+</default></profiles>
+</yandex>
+EOT
+
+fi